MHonArc 1.0.0 (MS-DOS beta release):
The source for this revision is MHONB1.ZIP, found on a German FTP
site: https://ftp.gwdg.de/pub/sgml/ifi.uio.no/DTD2HTML/MHONB1.ZIP.

Inside the zipfile is a diff file, MHONARC.DIF, that starts out with:

  This is a minor MS-DOS adaptation of MHonArc 1.0.0, released for beta testing.
  Steven Pacenka, sp17@cornell.edu, 16 November 1994

The rest of the diff file contains some minor changes that presumably
enable MHonArc 1.0.0 to run on MS-DOS.

We've placed this release first in the Git revision history because
that way you can see the MS-DOS diff (albeit reversed) by diffing this
revision with the upstream MHonArc 1.0.0 revision after it, and the
difference from upstream MHonArc 1.0.0 to the next release will be
correct by default (instead of being interfered with by the insertion
of the MS-DOS port).  Another solution, and perhaps a more formally
correct one, would have been to put the MS-DOS port on its own branch
based from the 1.0.0 upstream revision.  However, that seemed likely
lead to the MS-DOS port not getting seen -- Conservatory are usually a
linear sequence of releases, and people might not look for branches.